fix(spec): add DB_Store mock to spec_helper, fix cross-test contamination
The slackbot spec's inline DB_Store mock was clobbering the module
for qstrike specs when run in the full suite. Move the mock to
spec_helper following the established pattern (in-memory store with
reset!) so all specs that need DB_Store get a consistent, isolated mock.
